Стабильная версия с всеми вынесеными тестовыми классами
